main_for_service now has independent tests.
