+++ flocker/control/test/__init__.py
@@ -1,0 +1,6 @@
+# Copyright Hybrid Logic Ltd.  See LICENSE file for details.
+
+"""
+Tests for ``flocker.control``.
+"""
+
