+++ flocker/node/test/test_script.py
@@ -6,11 +6,6 @@ Tests for :module:`flocker.node.script`.
 
 from StringIO import StringIO
 
-from zope.interface import implementer
-
-from twisted.test.proto_helpers import MemoryReactor
-from twisted.internet.interfaces import IReactorCore
-from twisted.internet.defer import Deferred
 from twisted.trial.unittest import SynchronousTestCase
 from twisted.python.usage import UsageError
 from twisted.python.filepath import FilePath
@@ -401,137 +396,8 @@ class ReportStateScriptMainTests(SynchronousTestCase):
         self.assertEqual(safe_load(content.getvalue()), expected)
 
 
-# TODO: This should be provided by Twisted (also it should be more complete
-# instead of 1/3rd done).
-from twisted.internet.base import _ThreePhaseEvent
 
-
-@implementer(IReactorCore)
-class MemoryCoreReactor(MemoryReactor):
-    """
-    Fake reactor with listenTCP and just enough of an implementation of
-    IReactorCore to pass to ``_main_for_service`` in the unit tests.
-    """
-    def __init__(self):
-        MemoryReactor.__init__(self)
-        self._triggers = {}
-
-    def addSystemEventTrigger(self, phase, eventType, callable, *args, **kw):
-        event = self._triggers.setdefault(eventType, _ThreePhaseEvent())
-        event.addTrigger(phase, callable, *args, **kw)
-        # removeSystemEventTrigger isn't implemented so the return value here
-        # isn't useful.
-        return object()
-
-    def fireSystemEvent(self, eventType):
-        event = self._triggers.get(eventType)
-        if event is not None:
-            event.fireEvent()
-
-
-class AsyncStopService(Service):
-    """
-    An ``IService`` implementation which can return an unfired ``Deferred``
-    from its ``stopService`` method.
-
-    :ivar Deferred stop_result: The object to return from ``stopService``.
-        ``AsyncStopService`` won't do anything more than return it.  If it is
-        ever going to fire, some external code is responsible for firing it.
-    """
-    def __init__(self, stop_result):
-        self.stop_result = stop_result
-
-    def stopService(self):
-        Service.stopService(self)
-        return self.stop_result
-
-
-class ServeScriptMainTests(SynchronousTestCase):
-    """
-    Tests for ``ServeScript.main``.
-    """
-    def setUp(self):
-        self.reactor = MemoryCoreReactor()
-        self.service = Service()
-        self.script = ServeScript()
-
-    def main(self, reactor, service):
-        options = ServeOptions()
-        options.parseOptions([])
-        return self.script.main(reactor, options, service)
-
-    def _shutdown_reactor(self, reactor):
-        """
-        Simulate reactor shutdown.
-
-        :param IReactorCore reactor: The reactor to shut down.
-        """
-        reactor.fireSystemEvent("shutdown")
-
-    def test_starts_service(self):
-        """
-        ``ServeScript.main`` accepts an ``IService`` provider and starts it.
-        """
-        self.main(self.reactor, self.service)
-        self.assertTrue(
-            self.service.running, "The service should have been started.")
-
-    def test_returns_unfired_deferred(self):
-        """
-        ``ServeScript.main`` returns a ``Deferred`` which has not fired.
-        """
-        result = self.main(self.reactor, self.service)
-        self.assertNoResult(result)
-
-    def test_fire_on_stop(self):
-        """
-        The ``Deferred`` returned by ``ServeScript.main`` fires with ``None``
-        when the reactor is stopped.
-        """
-        result = self.main(self.reactor, self.service)
-        self._shutdown_reactor(self.reactor)
-        self.assertIs(None, self.successResultOf(result))
-
-    def test_stops_service(self):
-        """
-        When the reactor is stopped, ``ServeScript.main`` stops the service it
-        was called with.
-        """
-        self.main(self.reactor, self.service)
-        self._shutdown_reactor(self.reactor)
-        self.assertFalse(
-            self.service.running, "The service should have been stopped.")
-
-    def test_wait_for_service_stop(self):
-        """
-        The ``Deferred`` returned by ``ServeScript.main`` does not fire before
-        the ``Deferred`` returned by the service's ``stopService`` method
-        fires.
-        """
-        result = self.main(self.reactor, AsyncStopService(Deferred()))
-        self._shutdown_reactor(self.reactor)
-        self.assertNoResult(result)
-
-    def test_fire_after_service_stop(self):
-        """
-        The ``Deferred`` returned by ``ServeScript.main`` fires once the
-        ``Deferred`` returned by the service's ``stopService`` method fires.
-        """
-        async = Deferred()
-        result = self.main(self.reactor, AsyncStopService(async))
-        self._shutdown_reactor(self.reactor)
-        async.callback(None)
-        self.assertIs(None, self.successResultOf(result))
-
-    def test_starts_http_api_server(self):
-        """
-        ``ServeScript.main`` starts a HTTP server on the given port.
-        """
-        self.script.main(self.reactor, {"port": 8001}, self.service)
-        server = self.reactor.tcpServers[0]
-        port = server[0]
-        factory = server[1].__class__
-        self.assertEqual((port, factory), (8001, Site))
+class XXX:
 
 
 class StandardServeOptionsTests(
